fix: wing-playground build fails intermittently (#1481)
wing-playground was not using `file:` in its package json. I think this caused it to not do a symlink, but rather a full "install" of wingsdk.

While trying to solve this, I was updating several package lockfiles (`npm audit fix`) so I figured I would keep this in the PR cause why not.

Hopefully fixes #1479

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.