chore: limit parallelism for root `npm install` (#1483)
Trying to figure out ways to address #1479

Still not clear on root cause, but I have a hunch it's related to the npm install process. So with that, I made it so that the root "install everything" script doesn't run all installs at once to hopefully avoid whatever corruption is happening. Frankly this is probably a decent idea anyways, with no throttling we were surely hitting some io (net or fs) wall anyways.

Also, I saw that the lockfile for wing-playground was missing a `name` for wingsdk. Not sure if it matters, since I don't think module resolution cares about lockfiles, but hey why not make sure that's in here.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.