chore: pin node/npm version for build workflow (#1496)
Ok so during npm install symlinks were not being created where required. I believe this is some sort of regression with a new minor/patch version of node or npm. Using volta we already pin the node version locally, so it makes sense to do it in the workflow as well.

Removed any explicit node versioning where it is unnecessary.

Also combined all of the build into 1 command, because it shows the power of nx.

Fixes #1479 🤞 

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.