fix(sdk): reduce default lambda timeout to 30 seconds (#1508)
p0 - hello.w terraform plan doesn't work.

The queue visibility timeout is somehow set by default to 30 seconds, while the lambda timeout is configured to 60 seconds by us.
To fix this as quickly as possible we simply reduce the lambda timeout to 30 seconds as well. In a later PR we will increase both to 60 seconds by default.

Fixes #1507  

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.