fix: unclear compiler errors and lack of error snapshot testing (#1499)
Ever since wingc stopped being a binary, calling `compile` doesn't cause an error unless there's a panic. This means that execution isn't stopped in the CLI for regular errors but the rest of the cli expects the preflight js to exist, causing it to throw trying to read the file.

To help verify this, added snapshot testing for the `invalid` tests to hangar.

Fixes #926

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.