feat: isomorphic wasi (#1571)
### Current State

wingc is currently consumed in two ways:
- in the cli, via node's `wasi` modules
- in the playground, via https://github.com/wasmerio/wasmer-js

This is ok, but there are several downsides:
- Some code duplication with initialization and invocation of wingc
- wasmer-js is frankly not well maintained
- node's `wasi` only works in node
  - not a big deal, but prevented usage in other node-compatible runtimes or https://webcontainers.io/
- node `wasi` required node18 or newer and more importantly requires the use of an experimental flag

### New Change

I noticed that [bun](https://bun.sh/) recently started using [wasi-js](https://www.npmjs.com/package/wasi-js) for wasi. This library is a better-maintained fork of wasmer-js. It's part of https://cowasm.org/ which is a super cool project.

So I switched the CLI to use this library to load wingc, and made the playground consume the cli as a library.

## TODO

- [x] Windows?
- [x] Node16?
- [x] Docs


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.