fix(compiler): capture fields of resource with no inflight methods (#1565)
This was a rabbit hole :rabbit2: :hole: thanks @Chriscbr 

The problem is that when _capture.rs_ wants to capture a resource is iterates over all its inflight methods and creates a list of the methods that are (potentially) used by the inflight closure. This is required to pass to the sdk later on. If there are no inflight methods we just "skip" capturing which is a bug.

The quick fix is that if there are any capturable fields in the resource we just add the resource as an immutable data capture (in addition to a resource capture, see `CaptureKind`). But this caused the captured resource to be handled twice in the sdk's `makeHandler`. And then we ran into ordering issues. So the quick fix here was to change the order of some code in the sdk so the double capture won't cause issues.

### note about a possibly better implementation

As @Chriscbr mentioned in #1559, another approach can be not to capture the resource when accessing its immutable fields but instead capture the field itself. But this approach leads us to needing to capture "expressions" because in
```wing
new cloud.Function(inflight ()=> {
  print(some_resource.inner_resource.field);
});
```
capturing the nested identifier `some_resource.inner_resource` requires capturing a preflight expression and turning it into an inflight symbol. This introduces the ability to capture any preflight expression (as long as it evaluates to immutable data). Something like:
```wing
new cloud.Function(inflight ()=> {
  print(some_resource.preflight_method_that_returns_a_resource().field);
});
```
This mechanism can be very cool but requires the capture mechanism and jsifier to do some trickery when jsifying the inflight closure:
1) replacing preflight expressions with a generated symbol.
2) adding the generated symbol to the bindings.
3) initializing the binding with the JS code that'll produce the result of the expression.

All this seems less critical for now, so we'll just cpature the entire resource and rely on the fact that its inner fields are "lifted".


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.