chore(compiler): visitor trait (#1574)
Introduces a new Visit trait in the compiler that allows us to use the visitor pattern. The visitor pattern makes it easy to add new compiler passes since it handles the logic of traversing the AST tree for you, only requiring you to specify implementations of methods for the kinds of AST nodes you're interested in.

I used the trait to clean up `capture.rs` and remove some boilerplate.

No user-impacting changes.

Closes #1570 

Misc changes:
- renamed `args` to `arg_list` to be consistent with the struct name, and the identically-named field in `ExprKind::New`
- renamed `Expr::Binary.lexp` and `rexp` to `left` and `right` (copilot suggested it)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.