feat(compiler): add exponent support (#1654)
Implement exponent mentioned: https://github.com/winglang/wing/issues/470.

NOTE: exponentiation operator precedence was not defined in the language spec, so I followed the JavaScript spec (considering we are jsifying them), putting it in between Parentheses and Unary minus. I inserted it before `NIL_COALESCING` in PREC, not sure if it would be an issue. LMK

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.