feat: bring jsii enums (#1661)
Was curious to see if `projen` was bringable. It was just missing enum support.

Fixes #1484

While doing that, I noticed it was strange that `find_type` didn't use the lovely FQN struct. It makes more sense, plus getting rid of the iterator in that function is probably a nice optimization. 
Most of the changes in this PR are to support this move.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.