chore: replace polycons with a `new()` factory method
Since we are optimizing the SDK to be used through Wing, we can remove the dependency on `polycons` and clean up much of the boilerplate needed in the base classes.

This refactor removes the dependency on `polycons` by altering the compiler to emit a call to `this.node.root.new(my_resource_fqn, ...)` instead of `new MyResource(...)`. This is only done for resources for which we actually have an FQN, currently only types imported from JSII. For resources without an FQN, we continue to emit normal `new`.

The construct returned by `this.node.root` is the `App`, which is target-specific and therefore can determine the concrete behavior based on the target. In the future we will be able to allow plugins to customize further.

Types are now identified for instantiation using their JSII FQNs (e.g. `@winglang/sdk.cloud.Bucket`), but in order to maintain backwards compatibility with the Wing Console, the old type naming scheme (e.g. `wingsdk.cloud.Bucket`) is still used in the simulator manifest. We will change this in a subsequent commit in coordination with the Wing Console.

This change allowed removing most of the boilerplate hackery required to implement polycons: most `XxxBase` classes were deleted as well as the phantom polycons classes. Yey!

To make it easier to create `cloud.Xxx` resources in the SDK code (either when resources use each other or in tests), each resource class has a static method called `cloud.Xxx.newXxx()` which basically calls `this.root.new()` with the resource's FQN. This is how you would define a bucket:

    import { Bucket } from "@winglang/sdk";
    Bucket.newBucket(scope, id, props);

Yes, this is not ideal, but it's also not horrible and