feat(compiler): static support (#1662)
See #107

also in this PR:
* de-dup diagnostics by storing them in an `IndexSet` instead of a
`Vec`.
* Added a new AST reference type `TypeMember` for referencing enum
fields and class statics. Currently the grammar has no way of
distinguishing between a nested reference and a type property reference.
So during type-checking I do a (unsafe) convert in the AST tree from one
type to the other. If we change the grammar to distinguish between the
two we can avoid the unsafe here.
* Changed `Reference::NestedReferece` to `Reference::InstanceMember` to
match nicely with the new `Reference::TypeMember`.

**Not in this PR**
  - Support static field initialization (#1668)
  - Support accessing preflight static stuff inflight (#1669)

*By submitting this pull request, I confirm that my contribution is made
under the terms of the [Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

---------

Co-authored-by: Chris Rybicki <chrisr@monada.co>