chore: replace polycons with a `new()` factory method (#1681)
Since we are optimizing the SDK to be used through Wing, we can remove the dependency on `polycons` and clean up much of the boilerplate needed in the base classes. This is the first step towards #1365.

To do this, we've changed the compiler to emit a call to:

```js
this.node.root.new(my_fqn, MyResource, scope, id, ...args)
```

instead of:

```js
new MyResource(scope, id, ...args)
```

We do this only for resources for which we actually have an FQN, currently only those imported from JSII. For resources without an FQN (such as user-defined types), we continue to emit normal `new`.

The node returned by `this.node.root` is our `App`, which is target-specific. Therefore, it can determine the concrete behavior based on the target. In the future we will be able to allow plugins to customize further. For now, I removed the `customFactory` option from app as it is not being used and will be replaced with a hook.

The `new()` factory has the following signature:

```js
new(fqn, ctor, scope, id, ...args)
```

Where `fqn` is the JSII FQN of the type we wish to create (e.g. `@winglang/sdk.cloud.Bucket`). `ctor` is the constructor object (e.g. `Bucket`) and the rest are the standard construct arguments.

The `ctor` is passed in order to allow the `core.App` base implementation to fall back to  simply instantiate the type requested (e.g. when defining raw CDKTF resources). If `ctor` is `undefined`, it implies that the resource is an *abstract type* and therefore must be resolved by the app. If it is not resolved, this is the error:

```
Unable to create an instance of abstract type "@winglang\/sdk.cloud.Schedule"
```

Types are now identified by their JSII FQNs, but in order to maintain backwards compatibility with the Wing Console, the legacy type naming scheme (e.g. `wingsdk.cloud.Bucket`) is still used for the simulator manifest. We will change this in a subsequent commit in coordination with the Wing Console.

This change allowed removing most of the boilerplate hackery required to implement polycons: most `XxxBase` classes were deleted as well as the phantom polycons classes. Yey!

To make it easier to create `cloud.Xxx` resources in the SDK code (either when resources use each other or in tests), each resource class has a static method called `cloud.Xxx.newXxx()` which basically calls `this.root.new()` with the resource's FQN. This is how you would define a bucket:

```js
cloud.Bucket.newBucket(scope, "MyBucket");
```


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.