fix(sdk): some bucket keys are invalid (#1724)
Previously, many kinds of keys could not be used for objects in `cloud.Bucket`. This would result in runtime errors in the Wing simulator (even these keys are valid in AWS and other cloud providers).

Fixed the issue by storing objects in the file system according to a hashed version of their key, and storing the set of keys in-memory.

Related to #1676

Misc: cleaned up some redundant test snapshots

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.