feat(compiler): minimal capturing from inflight methods (#1682)
Analyze inflight methods to identify exactly which fields are being
referenced and which operations are performed on them. Previously, we
assumed all inflight methods access all fields across all their inflight
operations.

## Implementation notes

This is currently implemented during jsification. The function
`find_inflight_references()` iterates over all methods in a resource.
For each method it uses a `FieldReferenceVisitor` to analyze the method
and identify all references to `this.xxx`.

In some cases we will only fail at runtime. For example see
[`inflight_ref_resource_collection.w`](https://github.com/winglang/wing/blob/eladb/infer-refs/examples/tests/valid/inflight_ref_resource_collection.w)
and
[`inflight_ref_unknown_op.w`](https://github.com/winglang/wing/blob/eladb/infer-refs/examples/tests/valid/inflight_ref_unknown_op.w)
are both supposed to fail and require that explicit annotations are
added. This will be handled as part of #1610.

The `JSifier` now maintains a `diagnostics` list so it will be possible
to report errors (this turned it into a mutable creature).

We still want to bail out *before* jsification if there are any errors
in previous phases because otherwise, the jsifier cannot infer certain
types, etc.

Tests are prefixed with `inflight_ref_*` (both valid and invalid).

Co-author: @yoav-steinberg 


*By submitting this pull request, I confirm that my contribution is made
under the terms of the [Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

---------

Signed-off-by: mutator <mutator@winglang.io>
Co-authored-by: mutator <mutator@winglang.io>
Co-authored-by: Yoav Steinberg <yoav@monfort.co.il>