feat: import interfaces - attempt 2 (#1735)
Adding the interface importing support from #1684, this time without some of the extra changes made to SDK and the compiler's assembly loading logic. (#1684 had to be reverted because the changes made to the SDK caused the playground to stop working, and caused the getting started example to stop working in some environments https://winglang.slack.com/archives/C04BBDQUWQP/p1677784725607179).

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.