feat(sdk): add api resource tf-aws (#1285)
# Cloud.Api
## SDK Cloud Api
The SDK Cloud Api is a Resource that contains a OpenApi specification. The Specification is generated by adding route to a empty specification. The specification is then used while synth to Terraform to describe the AWS Api Gateway. Therefore each endpoint or route is represented in the OpenApi specification and has a x-amazon-apigateway-integration property that describes the integration with the Lambda function.

## AWS Api implementation
The AWS implementation of the Cloud.Apiresource is split into two parts. The exported class Api implements the SDK specification. It orchestrates the interaction with the "Parent class" ApiBase for example calling the OpenApi manipulation methods to add new routes. The second part of the AWS implementation is a Abstraction class called WingRestApi. This class should encapsulate all the Terraform specific things. This includes creating a OpenApi extension that glues together the route and the handling lambda function as well as setup the required IAM policies.
All route handler use the same inflight wrapper from api.onrequuest.inflight.ts to handle a inflight request.
This inflight handler maps the AWS ApiGateway event payload to the ApiREquest model from the SDK specification. The ApiRequest model is then passed to the handler function. The handler function returns a ApiResponse model which is then mapped to the AWS ApiGateway response payload.


*By submitting this pull request, I confirm that my contribution is made under the terms of the 
[Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.