chore(repo): revert to default number of columns for the contributors section (#1782)
Hey there, after committing I realised that contributor avatars didn't increase in size when the column number was decreased, leaving a huge gap on the right hand size (as seen in the screenshot below). I've reverted to the default number of columns (12) which will fill up the readme file from left to right. 

![Screenshot 2023-03-11 at 11 40 42](https://user-images.githubusercontent.com/38757612/224482548-19f11d62-019b-4447-8e7e-89a7b6132633.png)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.