feat(sdk): adding bucket public url (#1750)
xFixing this issue: https://github.com/winglang/wing/issues/1382

Implementation was written and manually tested for the sim and tf-aws.  not yet implemented on tf-azure.
jest tests were added for sim and tf-aws: 
- in a case the bucket isn't public OR the key does not exist- throws an error
- other wise returns the public URL (or the location of the file in local env).

I could change the code easily to be able to deploy and serve this static ["website"](https://cloud-bucket-c87175e7-20230306105442688700000003.s3.us-east-2.amazonaws.com/hello.html
)

by adding website configuration to the bucket and content type to the uploaded files (can be implemented as another resource as in [SDK spec](https://docs.winglang.io/reference/wingsdk-spec#website) or/and by adding an additional property to the bucket constructor, that is a string of the entry point)

I believe it can be used by web developers, especially those who are having their first steps creating a single-page/static application, bringing us a lot of supportive audience (I wish when I just started the deployment part was so easy!)


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.