fix: panic while analyzing captures in unknown calls/news (#1822)
When type checking and coming across a bad (_read_: typed as `any`) call/new reference, we did not traverse the arguments. That means we never gave them type info. 
The capturing process (and lsp) assumes all expressions in the AST have a type, so it panics.

Also re-enabled vscode's capture diagnostics, because this bug is why I removed them in the first place.

Fixes #1747
Fixes #1647
Fixes #1802

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.