feat: better captures in resources
Rewrite the algorithm which analyzes the expressions captured by inflight methods so that it
is able to identify more cases and emit errors when captures cannot be qualified (i.e. a resource
is captured but we cannot determine which operations are performed on it without static analysis).

Merge all `inflight_ref_resource_*` valid tests into `resource_captures` and all negative tests as well.