fix: incomplete identifiers cause cascading errors at the end of a scope (#1844)
### Huh?
In the grammar, a "reference" by itself is also valid expression. However, a lot of expressions can also start with a reference (or reference-like, in the case of custom types). References can also contain an expression themselves.
Long story short, these properties together made references very greedy and they often gobbled up many tokens like "}" and ")" trying to find the next semicolon when the reference is incomplete. Not sure if this is a tree-sitter bug or just a badly written grammar.

#### Before
<img width="693" alt="image" src="https://user-images.githubusercontent.com/1237390/225927283-d8e8977b-9903-4977-af10-b84dbfc4be87.png">

### What changed?
The primary change here was to remove references as an expression, but still allow identifiers or nested identifiers. This should keep the same expressivity but removes an extra layer of recursion that was hurting parseability.

### How about now?

Incomplete identifiers now have more localized errors. This mostly matters for the vscode extension, but it should be useful for compilation errors in general

#### After
<img width="693" alt="image" src="https://user-images.githubusercontent.com/1237390/225927950-8c43a9c1-0b14-46f0-abb1-ca041ee99ae7.png">



*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.