feat: extern js functions (#1837)
I consider this a first pass at extern, it should at least satisfy the needs of story 15.

This includes the functionality to load node modules in that extern js file

Closes #478, per spec

The primary non-obvious limitation is that the js file referenced is moved, so users cannot use the assumption for its behavior (e.g. relative path references).

Controversial changes needed to support this:
- Removed the "_inflight" postfix from generated inflight clients.
  - AFAIK, this isn't needed for any functionality
  - This was mostly done as a hack to give inflight methods a way to reference static inflight within the same class
- Gave the infight sim function VM access to the parent's `require`
- Pass the entire process object to the synth context and add an additional env to use there
  - passing the env without the context wasn't living long enough and idk why

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.