feat!: improved performance for sim compilation (#1859)
Improve performance from *O(inflight)* to *O(1)* when compiling and
testing on the `sim` target by skipping cloud function bundling. As a
benchmark, compiling a program with a single cloud function took 6s and
now takes 1s. A program with 5 cloud functions took 24s and now also
takes 1s.

The `.wsim` format has changed. It is now a directory instead of a
tarball and includes the generated javascript files without bundling any
dependencies. The output directory is now fixed to be `target/xxx.wsim`
relative to the location of the Wing entrypoint file. This allows any
`require()` statements to resolve against the correct `node_modules`. In
the future we will consider some kind of self-contained bundle format,
but for now, we optimize for the local iteration case.

To avoid mistakes, we removed support for specifying the output
directory `--out` in the compiler. We can introduce later when we
implement bundling for the simulator (or forbid only for the simulator).

### Details

- **Compiler**: Changed `js_resolve_path` to produce relative paths
instead of absolute paths.
- **SDK**: Moved the bundling logic into a utility function under
`bundling.ts`
- **SDK**: Since `vitest` runs our tests in TypeScript and now our cloud
functions are not bundled, we need to compile the `on*.inflight.ts`
files to javascript so that they can be injected into the VM.
- **SDK**: Fixed #1855 by cloning `global` into the inflight VM.
- **SDK**: Cleaned up the API for `NodeJsCode`. There is actually no use
in `path`.
- **SDK**: Discovered that the simulator doesn't `await` the handler
when running cloud functions, so fixed it.

BREAKING CHANGE: **cli**: The `--outdir` switch has been removed and the
default location has been changed so that compilation artifacts are
always placed under `target/xxx.ttt` (where `xxx` is the base name of
the entrypoint and `ttt` is the target - for example,
`target/hello.tfaws`, `target/hello.wsim`).
* **sim**: The `.wsim` output is now a directory and not a file
(tarball), and must reside in a location where node module resolution
can find any external dependencies.



*By submitting this pull request, I confirm that my contribution is made
under the terms of the [Monada Contribution
License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.

---------

Signed-off-by: mutator <mutator@winglang.io>
Co-authored-by: mutator <mutator@winglang.io>
Co-authored-by: Chris Rybicki <chrisr@monada.co>