feat!: rename `print` to `log` (#1898)
Logs are a first class concern in building distributed cloud applications, so we would like to rename the global function `print` to `log` to make the intent of the API clearer (in both preflight and inflight phases). See https://github.com/winglang/wing/discussions/1772 for more details.

This PR updates the Language spec as the compiler to use the new function name.

BREAKING CHANGE: `print()` is now named `log()`

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.