feat: `wing test -t tf-aws` (#1926)
Adds support for testing apps against live cloud applications by running `wing test -t tf-aws`.

https://user-images.githubusercontent.com/5008987/229242303-53f65413-65bd-452c-b542-a56db0aa33a6.mov

In order to support this, a new `cloud.TestRunner` resource was added to the SDK. This resource doesn't represent any physical cloud resources itself, but it contains an inflight API for listing and running tests in the current app. The `cloud.TestRunner` resource is automatically added to the root of each app, just like `cloud.Logger`.

In order to ensure the tests behave the same in the cloud as in the simulator, when Wing source code is compiled as part of `wing test`, an isolated copy of the user's app is generated for each cloud test. (This is the same as how the simulator currently operates today). In each isolated environment, one test gets executed. The artifacts of this test variant of the application are synthesized into `target/test/`.

This model basically assumes the root of your application functions as the "setup" for each test. In the future, we can add support for sharing resources between tests through new extensions to the language or SDK.

Known issues / future work:
- `tf-gcp`, `tf-azure`, and `awscdk` targets are not supported
- Traces are not supported on the `tf-aws` target, so any logs generated by tests will not be emitted to the CLI if you test with this target.
- Loading spinner doesn't spin while compiling the app because compilation uses the main JavaScript thread
- Tests on AWS are currently invoked by making AWS API calls directly from your machine -- this means running `wing test` requires having a profile set up not only with permissions to deploy your app to AWS, but also to call any AWS Lambda functions
- We currently auto-approve all Terraform changes, and anything logged by Terraform is not emitted. Users may want to configure these.
- Test durations are not recorded.
- We've added `force_destroy` to all buckets so that integration tests can be cleaned up -- we will remove this once we have a mechanism to tell the SDK whether we are compiling in "test" mode or not.
- We currently create each test function in every test environment, even though we are only going to run one test per environment.

Closes #1368

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.