fix: extern has inconsistant module semantics (#1994)
The string used with extern has several limitations
- It must be a file
- The file must be a CJS module
- The file itself cannot have references to paths relative to itself (because it gets moved)

But if it's a CJS module, what's the point of the first restriction? A CJS module can take many forms as long as it can be imported by `require()`. This PR makes extern more aligned with that, which also fixes the last issue mentioned (since we are no longer copying the file).

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.