fix(compiler): remove redundant annotations (followup to #1988) (#2000)
In #1988 we add all the resource fields to all inflight method annotations in the resource, but all we really want is to make sure someone annotates these fields in at least 1 method (to resolve #1872).

In this PR I create a dummy `$init` inflight method which includes the required annotations so we can safely assume the host can instantiate a client to the resource with access to all inner fields.

I think this fix improves on the existing code because:
* There aren't any redundant annotations.
* We're clear that these annotations are specific for initializing the resource client, and not because of some explicit inflight field access.
* This resolves a bug in the existing code where it didn't filter the annotated fields based on whether they were capturable, reassingable, etc. We now rely on the existing code in `get_captures`.
* I was thinking that eventually we'll support custom inflight `init` methods for initializing a resource's client, these methods will also need annotations which can be added to the new `$init` annotation.


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.