fix(compiler): indentation in javascript output is a mess (#2045)
Migrate all relevant jsification code to `CodeMaker` so that now javascript output is nicely indented and much more human readable.

### Testing

I've manually reviewed snapshots and verified that only formatting changes were introduced. Use "hide whitespace" in GitHub to review.



### Implementation Notes

Any `jsify_xxx` function that is expected to return a block of code (statement, code block, etc) now returns a `CodeMaker` object instead of a string. Then, we can compose these objects together using `add_code()`.

The `jsify_expression()` function, and any other function that returns a part of a sentence still returns a string, but we have updated `CodeMaker` to handle multiline strings properly by breaking them down and adding each line to preserve the indentation of the parent.

When we have chained blocks (`if/elseif/else` or `try/catch/finally`), we do this Instead of putting the `else` in the same line as the closing brackets:

    if (cond1) {
      // code
    }
    else if (cond2) {
      // code
    }
    else {
      // code
    }

I think that's okay.

### Misc

This change uncovered a latent bug in how `try/finally` works. Previously we used to always generate a `catch` block, **even if the code did not have one**. This means that if we had only a `try/finally` block, then the exception will not be thrown. Nice!

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.