feat(sdk): add timeout handler to simulator queue (#2046)
Closes #165.

Console crash mentioned in #1213 should also be fixed with this PR, but I let this issue open because `retentionPeriod` still needs to be added to terraform queue implementations.

To be consistent I also changed the default timeout period in the simulator queue to 10s, since `@default Duration.fromSeconds(10)` is documented in base class `cloud/queue.ts`.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.