fix: unable to run extern-defined function in console (#2040)
Root Issue: The previous enhancement to extern relied on an environment variable `WING_PROJECT_DIR` to have an absolute path to where the .w file is. In running the Wing CLI this is set before doing anything so that's fine. However, when arbitrarily running the Simulator this was a hidden requirement. This is what caused the console to fail to run inflight externs.

Unfortunately, my optimistic quest to either avoid copying extern files or avoid absolute paths has failed. Unless we introduce a bundler to wingc or always run the bundler in preflight, we'll need to have an absolute path for extern files to ensure they're reachable everywhere.

Fixes #2032 (no console change needed)
Fixes #1877

CC @ainvoner

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.